requirements.txt
altgraph==0.17
astroid==2.3.3
auto-py-to-exe==2.6.6
autopep8==1.5.1
blinker==1.4
bottle==0.12.18
bottle-websocket==0.2.9
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cryptography==2.9
Eel==0.12.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Excel==0.0.7
Flask-Ext==0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
future==0.18.2
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
html2pdf==0.1
html5lib==1.0.1
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
jsonify==0.5
lazy-object-proxy==1.4.3
lml==0.0.9
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.770
mypy-extensions==0.4.3
mysql-connector==2.2.9
numpy==1.18.2
openpyxl==3.0.3
ordereddict==1.1
pdfkit==0.6.1
pefile==2019.4.18
Pillow==7.1.1
pprint==0.1
pycodestyle==2.5.0
pycparser==2.20
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.8
pylint==2.4.4
PyPDF2==1.26.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
pywin32-ctypes==0.2.0
pyxdg==0.26
PyYAML==5.3.1
pyyaxml==0.6.0
reportlab==3.5.42
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
tablib==1.1.0
texttable==1.6.2
Transcrypt==3.7.16
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.8
uuid==1.30
webencodings==0.5.1
Werkzeug==1.0.1
whichcraft==0.6.1
wkhtmltopdf==0.2
WooCommerce==2.1.1
wrapt==1.12.1
xhtml2pdf==0.2.4
xlrd==1.2.0
xlwt==1.3.0
xsmtplib==1.0.0.dev1